refactor(queue): decouple QueueModule from domain modules
Inverts the dependency direction so domain modules import QueueModule,
not the other way around. Eliminates all forwardRef cycles and fixes the
double-worker bug on the custom-briefing-generation queue.

- Move AUDIO_GENERATION_QUEUE provider from AudioModule → QueueModule
- Move ArticleProcessor from libs/queue/processors → ProcessorModule
- Move AudioGenerationProcessor from libs/queue/processors → AudioFilesModule
- Delete libs/queue/processors/ directory entirely
- Remove S3Module, AudioModule, and all domain forwardRef imports from QueueModule
