+++ libs/audio/audio.module.ts
@@ -1,34 +1,11 @@
-import { RedisModule, RedisService } from '@libs/redis';
+import { RedisModule } from '@libs/redis';
 import { Module } from '@nestjs/common';
-import { Queue } from 'bullmq';
-import { AUDIO_GENERATION_QUEUE } from '../queue/constants/queue.constants';
+import { QueueModule } from '../queue/queue.module';
 import { AudioJobService } from './services/audio-job.service';
 
 @Module({
-  imports: [
-    RedisModule,
-  ],
-  providers: [
-    {
-      provide: AUDIO_GENERATION_QUEUE,
-      useFactory: (redisService: RedisService) => {
-        return new Queue(AUDIO_GENERATION_QUEUE, {
-          connection: redisService.getClient(),
-          defaultJobOptions: {
-            attempts: 2,
-            backoff: { type: 'exponential', delay: 2000 },
-            removeOnComplete: { count: 100 },
-            removeOnFail: { count: 500 },
-          },
-        });
-      },
-      inject: [RedisService],
-    },
-    AudioJobService,
-  ],
-  exports: [
-    AUDIO_GENERATION_QUEUE,
-    AudioJobService,
-  ],
+  imports: [QueueModule, RedisModule],
+  providers: [AudioJobService],
+  exports: [AudioJobService],
 })
 export class AudioModule { }
